tests/test_dls.py
import pytest
from uninformed.dls import DFS_DLSLogic

class TestDLS:
    @pytest.fixture
    def dls_logic(self):
        # Create a mock canvas and other necessary objects
        mock_canvas = type('MockCanvas', (), {'create_text': lambda *args, **kwargs: None})()
        return DFS_DLSLogic(mock_canvas, lambda *args: None, lambda *args: None)

    def test_dls_simple_path(self, dls_logic):
        # Define a simple graph
        dls_logic.graph = {
            'A': ['B', 'C'],
            'B': ['D', 'E'],
            'C': ['F', 'G'],
            'D': ['H', 'I'],
            'E': [],
            'F': ['J', 'K'],
            'G': [],
            'H': ['L'],
            'I': ['M'],
            'J': ['N'],
            'K': [],
            'L': [],
            'M': [],
            'N': []
        }

        # Run DLS
        path = dls_logic.dls('A', 'F')

        # Assert the correct path is found
        assert path == ['A', 'C', 'F']

    def test_dls_no_path(self, dls_logic):
        # Define a graph with no path to the goal
        dls_logic.graph = {
            'A': ['B', 'C'],
            'B': ['D', 'E'],
            'C': ['F', 'G'],
            'D': ['H', 'I'],
            'E': [],
            'F': ['J', 'K'],
            'G': [],
            'H': ['L'],
            'I': ['M'],
            'J': ['N'],
            'K': [],
            'L': [],
            'M': [],
            'N': []
        }

        # Run DLS
        path = dls_logic.dls('A', 'O')

        # Assert that no path is found
        assert path is None

    def test_dls_goal_not_in_graph(self, dls_logic):
        # Define a graph where the goal is not present
        dls_logic.graph = {
            'A': ['B', 'C'],
            'B': ['D', 'E'],
            'C': ['F', 'G'],
            'D': ['H', 'I'],
            'E': [],
            'F': ['J', 'K'],
            'G': [],
            'H': ['L'],
            'I': ['M'],
            'J': ['N'],
            'K': [],
            'L': [],
            'M': [],
            'N': []
        }
        # Run DLS
        path = dls_logic.dls('A', 'Z')

        # Assert that no path is found
        assert path is None

    # Add more test cases as needed
